requirements.txt
bs4
requests
gspread
python-dotenv
pandas
transformers
openpyxl
torch
accelerate
sympy==1.12
pebble 
timeout-decorator 
attrdict
crawl4ai
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
google-custom-search